Option to disable the automatic purge of entities (#1933)
(cherry picked from commit d0180f05d5be58171702e57327f421be49c55ac8)

+++ src/main/java/com/onarandombox/MultiverseCore/MVWorld.java
@@ -282,7 +282,9 @@ public class MVWorld implements MultiverseWorld {
                 }
                 world.setSpawnFlags(allowMonsters, allowAnimals);
             }
-            plugin.getMVWorldManager().getTheWorldPurger().purgeWorld(MVWorld.this);
+            if (MultiverseCoreConfiguration.getInstance().isAutoPurgeEnabled()) {
+                plugin.getMVWorldManager().getTheWorldPurger().purgeWorld(MVWorld.this);
+            }
             return super.validateChange(property, newValue, oldValue, object);
         }
     }
